# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CourseItemKind = typeof CourseItemKind[keyof typeof CourseItemKind];


export const CourseItemKind = {
  LESSON: 'LESSON',
  QUESTION: 'QUESTION',
  SCENARIO: 'SCENARIO',
} as const;

/**
 * JSON value stored with an item (for example, question options or an answer).
 */
export type JsonValue = string | number | boolean | { [key: string]: unknown } | unknown[];

export interface Citation {
  citation?: string;
  pubId?: string;
  section?: string;
  para?: string;
  page?: string;
  [key: string]: unknown;
 }

export type CourseItemStatus = typeof CourseItemStatus[keyof typeof CourseItemStatus];


export const CourseItemStatus = {
  APPROVED: 'APPROVED',
} as const;

export interface CourseItem {
  id: string;
  kind: CourseItemKind;
  stem: string;
  options: JsonValue | null;
  answer: JsonValue | null;
  rationale: string | null;
  citation: Citation | null;
  support: number | null;
  status: CourseItemStatus;
}

export interface Section {
  id: string;
  courseId: string;
  title: string;
  order: number;
  items: CourseItem[];
}

export interface Course {
  id: string;
  title: string;
  sourceId: string;
  createdAt: string;
  sections: Section[];
}

export interface CourseListResponse {
  courses: Course[];
}

export interface CourseResponse {
  course: Course;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}