# lib/api-zod/src/generated/types/courseItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Citation } from './citation';
import type { CourseItemKind } from './courseItemKind';
import type { CourseItemStatus } from './courseItemStatus';
import type { JsonValue } from './jsonValue';

export interface CourseItem {
  id: string;
  kind: CourseItemKind;
  stem: string;
  options: JsonValue | null;
  answer: JsonValue | null;
  rationale: string | null;
  citation: Citation | null;
  support: number | null;
  status: CourseItemStatus;
}