# lib/api-zod/src/generated/types/courseListResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Course } from './course';

export interface CourseListResponse {
  courses: Course[];
}