# lib/api-zod/src/generated/types/courseResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Course } from './course';

export interface CourseResponse {
  course: Course;
}