# lib/api-zod/src/generated/types/section.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CourseItem } from './courseItem';

export interface Section {
  id: string;
  courseId: string;
  title: string;
  order: number;
  items: CourseItem[];
}